Redirect unmatched paths to home.
The frontend routes of AngularJS are not recognized on the backend.
Until now this will result in a 404 page.
If you go to such frontend page (e.g. /signup) and refresh the page,
you would see a 404 on the same URI where you have just seen some content.

When redirecting to the homepage so AngularJS could load and present content,
we are also including the previous path in the fragment identifier.

AngularJS could recognize it and display the relevant view and controller.
If it does not recognize it as well it will "redirect" to its default route.